requirements.txt
docling==2.15.0
pypdfium2>=0.0.0
# pymupdf>=1.24.0  # used only by benchmarks; kept installed but not a production dep
ollama>=0.1.0
tiktoken>=0.5.0
psutil>=5.9.0
reportlab>=4.0.0
fpdf2>=2.7.0
tqdm>=4.66.0
numpy>=1.24.0
# faiss-cpu removed — replaced by sqlite-vec
scikit-learn>=1.3.0
pydantic>=2.7
pyyaml>=6.0
rich>=13.0
sqlite-vec>=0.1.0
sentence-transformers>=3.0
torch>=2.2